+++ packages/edition-slices/src/tiles/tile-c/styles/index.js
@@ -2,6 +2,7 @@ import { fonts, spacing } from "@times-components/styleguide";
 
 const styles = {
   container: {
+    height: "100%",
     paddingHorizontal: spacing(2),
     paddingTop: spacing(2)
   },